requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pdfplumber==0.11.4
pydantic==2.10.6
anthropic>=0.40.0
pymupdf>=1.24.0
python-dotenv>=1.0.0
supabase>=2.0.0
